Fix place where we were capturing block variable as higher local variable.
(That doesn't work in Ruby 1.9.3.)

Punctuated site building messages with puts, using puts space to try to guarantee that newline appears; also added flush to fake stdout's print command and threw in an extra puts at the start of running a command, trying to get console to clear as early as possible. Added dots to messages from autoglossary construction, as this can take a long time and user needs to know something is happening.